mamboms/requirements.txt
#
# Requirements for mamboms
#
# To see what requirements your current virtualenv is using type:
#
# $ pip freeze
#
#
# Requirements can be specified as:
#  - local files (eggs/ccg-python-build_v2_0.tar.gz)
#  - http        (http://ccg-python-build.googlecode.com/files/ccg-python-build_v2_0.tar.gz)
#  - eggs        (python-ldap>=2.3.13)
#
# http://www.pip-installer.org/en/latest/requirement-format.html
#
http://mango-py.googlecode.com/files/Mango-py-1.3.1-ccg-196.tar.gz
http://ccg-django-extras.googlecode.com/files/ccg-extras-0.1.3.tar.gz
http://ccg-django-extras.googlecode.com/files/ccg-auth-0.3.1.tar.gz
http://ccg-django-extras.googlecode.com/files/ccg-makoloader-0.2.3.tar.gz
Cython==0.12
matplotlib==1.0.1
django-picklefield==0.1.9
django-templatetag-sugar==0.1
pyparsing==1.5.6
virtualenv==1.6.4
wsgiref==0.1.2
python-memcached==1.44
Mako>=0.5.0
South>=0.7.3
django-extensions>=0.7.1
psycopg2==2.0.8
python-ldap==2.3.13
file:eggs/search_datastructures-0.1-py2.6-linux-x86_64.egg
